fix(code): reconsolidates input managers
